add workout and ai to nav bar, rearrange nav bar links, split workout page to component and page, remove console logs from server
